fix the bug in cardelli_ext and add a unit test

+++ src/lib/ext.cpp
@@ -158,7 +158,7 @@ double cardelli_ext(flt &oneFlt) {
   // computes the galactic extinction
   double dlbd = (lmax - lmin) / 400.;
   for (size_t i = 0; i < 402; i++) {
-    lextg = lmin + (i - 1) * dlbd;
+    lextg = lmin + double(i - 1) * dlbd;
     extg = cardelli_law(lextg);
     oneExt.add_element(lextg, extg, 2);
   }
